Folder and file restructure
