+++ src/clipio_editor/scripts/ClipioModule.js
