+++ src/main/ipcmain-channels/events/DeleteModule.js
@@ -16,4 +16,5 @@ class DeleteModule extends Event {
   }
 }
 
-module.exports = DeleteModule;
+module.exports = DeleteModule;  
+
