+++ src/main/ipcmain-channels/events/GetAppVersion.js
