+++ src/main/ipcmain-channels/events/GetGithubToken.js
