+++ src/main/ipcmain-channels/events/GetLocalModules.js
