+++ src/main/ipcmain-channels/events/GetModuleManifest.js
