+++ src/main/ipcmain-channels/events/GetModulePath.js
