+++ src/main/ipcmain-channels/events/Maximize.js
