+++ src/main/ipcmain-channels/events/Minimize.js
