+++ src/main/ipcmain-channels/events/Relaunch.js
