+++ src/main/ipcmain-channels/events/SetGithubToken.js
