+++ src/main/ipcmain-channels/events/SetLocalModules.js
