+++ src/main/ipcmain-channels/utils/resetModulesFolder.js
